% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plasmode.R
\name{plasmode}
\alias{plasmode}
\title{Plasmode simulation for randomly selected outcome.}
\usage{
plasmode(outcome_data, target_outcome = NULL, target_inc = NULL, rr = 2)
}
\arguments{
\item{outcome_data}{Data with patient id, treatment status, covariates, and 
outcomes, as returned by select_subsample().}

\item{target_outcome}{A user-specified outcome code.}

\item{target_inc}{A user-specified incidence rate.}

\item{rr}{Desired level of simulated relative risk.}
}
\value{
A resampled dataset with a simulated outcome enriched to the desired 
  relative risk.
}
\description{
Use plasmode simulation to enrich one outcome to the desired 
  level of relative risk.
  Specify exactly one of target_outcome (if you have pre-selected the outcome
  you wish to enrich) or target_inc (if you wish for the program to select a 
  random outcome close to the specified incidence rate).
}
