% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_strat.R
\name{ps_strat}
\alias{ps_strat}
\title{Propensity score stratification for Poisson probability model.}
\usage{
ps_strat(input_data, strat_args)
}
\arguments{
\item{input_data}{Data with patient id, treatment status, covariates, and 
outcomes, as returned by plasmode().}

\item{match_args}{List of arguments for matching algorithm as specified in 
strat_params.R.}
}
\value{
Observed and expected outcome counts for treated patients
}
\description{
Propensity score sratification and indirect stratification for 
  use in TBSS with Poisson probability model. Arguments are specified in the 
  file strat_params.R
}
