% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_subsample.R
\name{select_subsample}
\alias{select_subsample}
\title{Select random subsample of desired size for power calculation.}
\usage{
select_subsample(outcome_data, target_treated_n, target_comp_n)
}
\arguments{
\item{outcome_data}{Patient data as retured by get_data().}

\item{target_treated_n}{Number of treated patients to select.}

\item{target_comp_n}{Number of comparator patients to select.}
}
\value{
A dataset with the selected number of treated and comparator patients
}
\description{
Select desired treated and comparator patient group sizes
}
