% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbss_mask.R
\name{tbss_mask}
\alias{tbss_mask}
\title{TBSS binary detection flag for specified outcome.}
\usage{
tbss_mask(df, model = c("poisson", "bernoulli"), tree_obj, sig = 0.05)
}
\arguments{
\item{df}{Data as formatted by ps_match() (Bernoulli) or ps_strat() (Poisson).}

\item{model}{Probability model, either Poisson or Bernoulli}

\item{tree_obj}{Hierarchical tree with columns 'parent' and 'child' as in 
output from tree_builder()}

\item{sig}{Desired level of significance.}
}
\value{
A binary indicator if outcome of interest is a statistical alert
}
\description{
Runs TBSS with specified probability model, masking specific 
  results and returning 1/0 if outcome of interest is/is not detected
}
